requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
psycopg2==2.9.10; sys_platform == 'windows'
psycopg2-binary==2.9.10; sys_platform == 'linux' or sys_platform == "darwin"
pytz==2024.2
sqlparse==0.5.2
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-environ==0.11.2
django-filter==24.3
django-rest-passwordreset==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
pillow==11.0.0
pipenv==2024.4.0
python-decouple==3.8
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
Unidecode==1.3.8
uritemplate==4.1.1
drf-yasg==1.21.8
whitenoise==6.8.2
python-decouple==3.8
django-guardian==2.4.0
twilio
gunicorn